Updated to have correct analyte/titrant strengths.
Changed strong pKa and pKt values to -6. This is correct for a strong acid/base. ~= to the pKa for HCl

Put objects into a list and called the plot function on each. This made it easier to change the plotting function being run.

Reordered the objects to be sorted by analyte type, acidic or basic.

There are now 6 of each analyte titrations.