New derivative normalization function.
The new function is = x / (1+|x|)

The new function solves issue #9.

The zero point of the second derivative now stays put when scaled because the normalizing range of the new function only majorly affects values outside of the range [-1,1] rather than the range [0, 1].

At low point counts there is no longer a seemingly step wise step in the second derivative.