Choosing a hard set maximum and minimum means that a lot of extraneous volume points are being calculated.
The start pH is now the pH of the analyte, as though it were a strong analyte. This pH is calculated with 14 - log10(conc), if the compound is basic, and -log10(conc) if the compound is acidic.

The end pH follows the same principle, but using the titrant rather than the analyte.

This change has the effect of putting more analysis within the valid region of the titration, rather than in the extremes of the titration which are impossible to reach.